New version of the 'component.check' module, the terror of the programmer.

+++ Examples/TSP/NSEnumSwap.hpp
@@ -173,7 +173,7 @@ public:
 		if (k2 == ((int)rep.size()) - 1)
 			ak2 = 0;
 
-		int f = 0;
+		double f = 0;
 
 		if (k2 - k1 == 1) // special case, cities are near
 		{
